requirements.txt
# Core dependencies
langchain
langchain-community
langchain-huggingface
# Removed: langchain-openai
# Removed: openai
pinecone-client
faiss-cpu

# Local AI Models
transformers
torch
sentence-transformers
accelerate

# PDF processing
pypdf

# Web framework
fastapi
uvicorn[standard]
python-multipart

# Frontend
streamlit

# Utilities
python-dotenv
pydantic
pydantic-settings
requests
PyJWT

# Development & Testing
pytest
pytest-asyncio
httpx
black
flake8
